refactor(stella-cli): split config.rs by concern
crates/stella-cli/src/config.rs sat at exactly 1500 lines, the ungrandfathered
ceiling, so the next Config field with an ordinary doc comment would have
failed the file-size gate outright (#3566, and #4494's cli half). Split three
self-contained concerns into new siblings under config/, following the
pattern aux_credentials/listing/providers/reload already established:

- launcher_overrides.rs: the trusted-launcher agent_engine_config JSON,
  --upstream-pin, --allow-dir, and the interactive-credential-prompt latch.
  All of it is a fact about the invocation main establishes once, none of
  it depends on Config/ProviderConfig. Re-exported at the old paths
  (forbid_interactive_credentials/set_upstream_pin/set_allow_dirs stay
  pub(crate) since main.rs calls them via crate::config::*; the five
  internal accessors are private `use`s, plus a #[cfg(test)] use for the
  two names config/tests/trusted_engine.rs and config/tests/credential.rs
  reach through the `use super::*;` chain).
- report.rs: a second `impl Config` block (print_models/print_config/
  print_role_wiring) -- the same shape reload.rs already uses for the one
  mutation Config supports after construction. This is the complementary
  read-only half.
- discovery.rs: ConfiguredProvider + discover_configured_providers, moved
  whole rather than split apart -- resolving ONE provider (config.rs's own
  job) and enumerating EVERY currently-configured provider are different
  questions, even though discovery reuses config.rs's own credential chain
  (resolve_provider_key, still defined in config.rs and called via
  super::). Re-exported pub at the old crate::config:: path -- it is used
  from a dozen call sites across the crate (command_deck, model_catalog,
  profile, subagent, agent::engine, agent::goal, ...), all of which keep
  compiling unchanged.

config.rs itself: 1500 -> 1144 lines, real headroom rather than the
one-or-two-line kind the issue warned against. No fields moved (the six
struct-literal construction sites #3566 named are untouched), no baseline
entry added or needed, no behavior change -- purely a code move plus the
visibility/re-export plumbing to keep every existing call site and test
resolving at its old path.

Closes #3566
Refs #4494
